refactor(backend): extrair handlers de cards para arquivos separados
- Criar createCard.ts, updateCard.ts, deleteCard.ts, getCard.ts e moveCard.ts
- Refatorar cards.ts para usar os novos handlers
- Melhorar organização e manutenibilidade do código
